+++ postcss.config.js
@@ -9,6 +9,7 @@ module.exports = {
     'postcss-calc': {},
     'postcss-color-function': {},
     'postcss-flexbugs-fixes': {},
+    'postcss-selector-matches': {},
     autoprefixer: {},
   },
 };